Add resume to file download

+++ putio.py
@@ -9,6 +9,15 @@ from urllib import urlencode
 import requests
 import iso8601
 
+KB = 1024
+MB = 1024 * KB
+
+# Downloads are broken up into ranges of this size
+RANGE_SIZE = 10 * MB
+
+# Downloads are written and flushed after a chunk of this size
+CHUNK_SIZE = 8 * KB
+
 BASE_URL = 'https://api.put.io/v2'
 ACCESS_TOKEN_URL = 'https://api.put.io/v2/oauth2/access_token'
 AUTHENTICATION_URL = 'https://api.put.io/v2/oauth2/authenticate'
@@ -165,11 +174,11 @@ class _File(_BaseResource):
         """List the files under directory."""
         return self.list(parent_id=self.id)
 
-    def download(self, dest='.', delete_after_download=False):
+    def download(self, dest='.', delete_after_download=False, range_size=RANGE_SIZE, chunk_size=CHUNK_SIZE):
         if self.content_type == 'application/x-directory':
             self._download_directory(dest, delete_after_download)
         else:
-            self._download_file(dest, delete_after_download)
+            self._download_file(dest, delete_after_download, range_size, chunk_size)
 
     def _download_directory(self, dest='.', delete_after_download=False):
         name = self.name
@@ -186,24 +195,40 @@ class _File(_BaseResource):
         if delete_after_download:
             self.delete()
 
-    def _download_file(self, dest='.', delete_after_download=False):
-        response = self.client.request(
-            '/files/%s/download' % self.id, raw=True, stream=True)
+    def _download_file(self, dest='.', delete_after_download=False, range_size=RANGE_SIZE, chunk_size=CHUNK_SIZE):
+        filename = self.name.strip('"')
+        filepath = os.path.join(dest, filename)
 
-        filename = re.match(
-            'attachment; filename=(.*)',
-            response.headers['content-disposition']).groups()[0]
-        # If file name has spaces, it must have quotes around.
-        filename = filename.strip('"')
+        if os.path.exists(filepath):
+            first_byte = os.path.getsize(filepath)
+        else:
+            first_byte = 0
 
-        with open(os.path.join(dest, filename), 'wb') as f:
-            for chunk in response.iter_content(chunk_size=1024):
-                if chunk:  # filter out keep-alive new chunks
-                    f.write(chunk)
-                    f.flush()
+        logging.debug('file %s is currently %d, should be %d' % (filepath, first_byte, self.size))
 
-        if delete_after_download:
-            self.delete()
+        with open(filepath, 'ab') as f:
+            # Split up file into blocks of RANGE_SIZE each
+            while first_byte < self.size:
+                if first_byte + range_size < self.size:
+                    last_byte = first_byte + range_size
+                else:
+                    last_byte = self.size
+
+                logging.debug('download range %d - %d' % (first_byte, last_byte))
+
+                headers = { 'Range': 'bytes=%d-%d' % (first_byte, last_byte) }
+                response = self.client.request('/files/%s/download' % self.id, headers=headers, raw=True, stream=True)
+
+                for chunk in response.iter_content(chunk_size=chunk_size):
+                    if chunk:  # filter out keep-alive new chunks
+                        f.write(chunk)
+                        f.flush()
+
+                first_byte = last_byte + 1
+    
+        if self.size == os.path.getsize(filepath):
+            if delete_after_download:
+                self.delete()
 
     def delete(self):
         return self.client.request('/files/delete', method='POST',
